+++ package.json
@@ -31,14 +31,14 @@
     "prettier": "^3.5.3",
     "prisma": "^6.6.0",
     "typescript": "^5.8.2",
-    "typescript-eslint": "^8.28.0"
-  },
-  "dependencies": {
+    "typescript-eslint": "^8.28.0",
     "@types/cookie-parser": "^1.4.8",
     "@types/cors": "^2.8.17",
     "@types/fluent-ffmpeg": "^2.1.27",
     "@types/jsonwebtoken": "^9.0.9",
-    "@types/multer": "^1.4.12",
+    "@types/multer": "^1.4.12"
+  },
+  "dependencies": {
     "argon2": "^0.41.1",
     "chalk": "^5.4.1",
     "cookie-parser": "^1.4.7",
