+++ src/api/v1/routes/user.ts
@@ -1,15 +1,18 @@
 import express from "express";
-import { checkLogin } from "../middleware/authCheck.js";
-import { UserController } from "../controller/user/userController.js";
-import { uploadFields } from "../middleware/multer.js";
 import { checkSchema } from "express-validator";
+import { UserController } from "../controller/user/userController.js";
+import { VideoController } from "../controller/user/videoController.js";
+import { checkLogin } from "../middleware/authCheck.js";
+import { uploadFields } from "../middleware/multer.js";
 import { uploadVideoSchema } from "../validation/user.js";
 
 const router = express.Router();
 
 // [OPEN Routes]
-router.get("/feed", UserController.getFeedData);
-router.get("/watch/:uuid", UserController.getVideo);
+router.get("/:username/video/:filename", VideoController.videoStream);
+router.get("/:username/thumbnail/:filename", VideoController.getThumbnail);
+router.get("/feed", VideoController.getFeedData);
+router.get("/watch/:uuid", VideoController.getVideo);
 
 // [PROTECTED Routes]
 router.get("/user/:id", checkLogin, UserController.getUser);
@@ -18,7 +21,7 @@ router.post(
   checkLogin,
   uploadFields,
   checkSchema(uploadVideoSchema),
-  UserController.uploadVideo
+  VideoController.uploadVideo
 );
 // router.patch("/user/:id", checkLogin, UserController.updateUser);
